fix flair trainer

+++ src/flair_trainer.py
@@ -67,7 +67,7 @@ class FlairTrainModel:
 
     def compute_w(self, train_loader : DataLoader, num_classes):
         _, targets = train_loader.dataset.data.tensors
-        t = targets.numpy()
+        t = targets.int().numpy()
         freq = np.bincount(t)
         return torch.Tensor(1 - (freq/freq.sum()))
